\name{datasetDrosophila}
\docType{data}
\alias{datasetDrosophila}
\title{Genome-wide RNAi screen of cell viability in Drosophila Kc167 cells by Boutros et al.}
\description{M. Boutros et al., Genome-wide RNAi analysis of growth and viability in Drosophila cells, Science, 303(5659):832-835, 2004. 3, 18}
\usage{datasetDrosophila}
\format{see \code{\link{generateDatasetFile}} for details}
\keyword{datasets}