\name{gseaAnalysis}
\alias{gseaAnalysis}

\title{ Perform a GSEA analysis of a list of genes }
\description{
  Performs a GSEA analysis of a list of genes using the package \code{ topGO} (see References).
}
\usage{
gseaAnalysis(hitVector, whichOnto)
}

\arguments{
  \item{hitVector}{ a named hit vector as generated by \code{\link{hitselectionZscore}} or \code{\link{hitselectionPval}} }
  \item{whichOnto}{ One of the three GO ontologies: \code{ "biological_process"}, \code{ "molecular_function"} or \code{ "cellular_component"} }
}

\value{
A table containing the enriched GO terms and their significance.
}
\references{ A. Alexa, J. Rahnenfuhrer and T. Lengauer. Improved scoring of functional groups from gene expression data by decorrelating GO graph structure. Bioinformatics, 22(13):1600-1607, 2006

Adrian Alexa and Jorg Rahnenfuhrer (2006). topGO: Enrichment analysis for Gene Ontology. R package version 1.4.0. }

\seealso{ \code{\link{GOannotate}} }
\examples{
data(scoredDataset1, package="RNAither")
data(pValVec1, package="RNAither")

##for details on the generation of pValVec1 and scoredDataset1, see \code{\link{Ttest}}

scoredHits1 <- hitselectionPval(scoredDataset1, pValVec1, "SigIntensity", "Hits1", 0.1, 
"GeneName", "pvalue_testfile1.txt")
hitVector1 <- scoredHits1[[2]]
gseaTable <- gseaAnalysis(hitVector1, "biological_process")

}
\keyword{ database }
\keyword{ datagen }