\name{headerDrosophila}
\docType{data}
\alias{headerDrosophila}
\title{the header of the genome-wide RNAi screen of cell viability in Drosophila Kc167 cells by Boutros et al.}
\description{M. Boutros et al., Genome-wide RNAi analysis of growth and viability in Drosophila cells, Science, 303(5659):832-835, 2004. 3, 18}
\usage{headerDrosophila}
\format{See \code{\link{generateDatasetFile}}}
\keyword{datasets}