\name{hitselectionPval}
\alias{hitselectionPval}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Selecting hits according to p-values }
\description{
  Selects significant genes according to their p-value.
}
\usage{
hitselectionPval(dataset, pValVec, col4val, col4sel, thresh, col4anno, file4hits)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{dataset}{ an R data frame generated with \code{\link{generateDatasetFile}} }
  \item{pValVec}{ a vector of p-values, as generated by one of the test functions \code{\link{Ttest}}, \code{\link{MannWhitney}} or \code{\link{RankProduct}} }
  \item{col4val}{ a character vector specifying a column of intensity values }
  \item{col4sel}{ a character vector specifying the name of the new dataset column where hits will be stored }
  \item{thresh}{ the threshold for the p-values, typically 0.05 }
  \item{col4anno}{ a character string specifying the name of the dataset column to be used to define the replicate, e.g. \code{"GeneName"} or \code{"Internal_GeneID"} }
  \item{file4hits}{ the name of the file to store the results in }
}
\details{
  If there are no p-values under the defined threshold \code{thresh}, the threshold is increased to \code{min(pvalvec)}.
  }
\value{

A list containing: 

  \item{dataset }{the dataset with an added column defining the hits in the form of a binary vector}
  \item{hitVector }{the binary vector itself}
  \item{replicaMatrix }{a matrix of replicates with corresponding values (as generated by \code{\link{generateReplicateMat}})}
  \item{thresh }{the threshold for the p-values}
  P-values and the intensity values for each siRNA are stored in a text output file.
}

\seealso{ \code{\link{hitselectionZscore}}, \code{\link{hitselectionZscorePval}} }
\examples{
data(scoredDataset1, package="RNAither")
data(pValVec1, package="RNAither")

##for details on the generation of pValVec1 and scoredDataset1, see \code{\link{Ttest}}

scoredHits1 <- hitselectionPval(scoredDataset1, pValVec1, "SigIntensity", "Pval_hits", 0.05, 
"GeneName", "pvalue_testfile1.txt")

newdataset <- scoredHits1[[1]]
hitvector <- scoredHits1[[2]]
}

\keyword{ arith }
