\name{hitselectionZscorePval}
\alias{hitselectionZscorePval}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Selecting hits according to ZScores and p-values }
\description{
  Selects significant genes according to their ZScore (summarized with the gene median) and p-values.
}
\usage{
hitselectionZscorePval(dataset, pValVec, col4zscore, col4sel, thresh, thresh2, 
flag2, col4anno, sumFunc, file4hits)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{dataset}{ an R data frame generated with \code{\link{generateDatasetFile}} }
  \item{pValVec}{ a vector of p-values, as generated by one of the test functions \code{\link{Ttest}}, \code{\link{MannWhitney}} or \code{\link{RankProduct}} }
  \item{col4zscore}{ a character vector specifying the name of the column containing the ZScores, usually \code{"SigIntensity"} }
  \item{col4sel}{ a character vector specifying the name of the new dataset column where hits will be stored }
  \item{thresh}{ the threshold for the ZScores. The interpretation depends on the choice of the parameter \code{ flag2}. }
  \item{thresh2}{ the threshold for the p-values }
  \item{flag2}{ 2 or -2. If 2 is chosen, all Zscores greater than or equal to \code{ thresh} are chosen. If -2 is chosen, all Zscores smaller than or equal to \code{ thresh} are chosen. }
  \item{col4anno}{ a character string specifying the name of the dataset column to be used to define the replicate, for example \code{ "GeneName"} or \code{ "Internal_GeneID"}. }
  \item{sumFunc}{ the function used to summarize ZScore values, e.g. \code{ mean} or \code{ median}. }
  \item{file4hits}{ the name of the file to store the results in }
}
\details{
  If there are no p-values under the defined threshold \code{ thresh2}, it is increased to \code{ min(pvalvec)}. 
  
  If \code{ flag2} == -2 and there are no ZScores under the defined threshold \code{ thresh}, it is increased to \code{ min(ZScores)}. 
  
  If \code{ flag2} == 2 and there are no ZScores over the defined threshold \code{ thresh}, it is increased to \code{ max(ZScores)}.
 
If there are not hits for the combined threshold of p-values and ZScores, the ZScore threshold is changed until there is a hit.
}
\value{
A list containing:

  \item{dataset }{the dataset with an added column defining the hits in the form of a binary vector}
  \item{hitVector }{the binary vector itself}
  \item{thresh }{the threshold for the ZScores}
  \item{thresh2 }{he threshold for the p-values}
ZScores and p-values are stored in a text output file.
}

\seealso{ \code{\link{hitselectionPval}}, \code{\link{hitselectionZscore}} }
\examples{
data(scoredDataset1, package="RNAither")
data(pValVec1, package="RNAither")

##for details on the generation of pValVec1 and scoredDataset1, see \code{\link{Ttest}}

scoredHits1 <- hitselectionZscorePval(scoredDataset1, pValVec1, "SigIntensity", 
"Zscore_pval_hits", -1.5, 0.05, -2, "GeneName", median, "Zscores_pvals_testfile1.txt")

newdataset <- scoredHits1[[1]]
hitvector <- scoredHits1[[2]]
}
\keyword{ arith }
