\name{incorporatepValVec}
\alias{incorporatepValVec}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Incorporate a vector of p-values into a dataset }
\description{
  Incorporates a vector of p-values into a dataset. Also works with a dataset containing values per well (non summarized), or with a hit vector.
}
\usage{
incorporatepValVec(dataset, pValVec, replicaMatrix, col4anno, colname4pval)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{dataset}{ an R data frame generated with \code{\link{generateDatasetFile}} }
  \item{pValVec}{ a vector of p-values }
  \item{replicaMatrix}{ a matrix of replicate values, as generated by \code{\link{generateReplicateMat}} }
  \item{col4anno}{ a character string specifying the name of the dataset column to be used to define the replicate, for example \code{ "GeneName"} or \code{ "Internal_GeneID"} }
  \item{colname4pval}{ a character string specifying the name of the dataset column the p-values will be stored in }
}

\value{
Returns the dataset with an added column of p-values.
}

\seealso{\code{\link{multTestAdjust}}, \code{\link{Ttest}} }
\examples{
data(exampleDataset, package="RNAither")

data(scoredDataset1, package="RNAither")
##scoredDataset1 already contains the p-value column
data(pValVec1, package="RNAither")

##for details on the generation of pValVec1 and scoredDataset1, see the example of the Ttest function linked above.

temp <- generateReplicateMat(dataset, 1, "Intensities", "SigIntensity",  "GeneName")
replicamatrix <- temp[[1]]
newdataset <- incorporatepValVec(dataset, pValVec1, replicamatrix, "GeneName", "pvals")
##newdataset and scoredDataset1 are now equivalent

}
\keyword{ manip }

