\name{RNAither-package}
\alias{RNAither-package}
\alias{RNAither}
\docType{package}
\title{
Statistical analysis of high-throughput RNAi screens
}
\description{
RNAither analyzes cell-based RNAi screens, and includes quality assessment, customizable normalization and statistical tests, leading to lists of significant genes and biological processes.
}
\details{
\tabular{ll}{
Package: \tab RNAither\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2008-07-20\cr
License: \tab Artistic License 2.0\cr
}

}
\author{
Nora Rieber and Lars Kaderali

Maintainer: Nora Rieber <Rieber Nor [at] gmx [dot] de>
}