\name{rnaither}
\alias{rnaither}

\title{ Wrapper function for full automated analysis }
\description{
  Performs a standard analysis of the data (quality and statistics) from a dataset file.
}
\usage{
rnaither(data, expname, excludeCellcounts="none", logtransform=FALSE, normalization=c("lowess","bscore"), test="ttest", scorethresh=2.0, pvalthresh=0.05, dogo=FALSE, outdir="results", layoutnames="NA", makeplots=TRUE, reorder=TRUE)               
}

\arguments{
  \item{data}{A data frame containing the experimental data to analyze. 
    Each row is corresponding to one well, with the following columns:
    \itemize{
    \item{Spotnumber}{ The position of the well on the plate}
    \item{Internal_GeneID}{ The ID of the siRNA}
    \item{GeneName}{ The gene name}
    \item{SpotType}{ Can be -1, 0, 1 or 2.
      Type -1 wells (e.g. emtpy wells, wells with poor quality) are not considered in subsequent analyses but are kept in the data set for the sake of completeness. 
      Type 0 wells correspond to negative controls, type 1 wells to positive controls. 
      Type 2 wells correspond to the standard data wells.}
    \item{SigIntensity}{ The signal intensity (channel 1)}
    \item{SDSIntensity}{ The standard deviation of the signal intensity, if available}
    \item{Background}{ The background per well, if available}
    \item{LabtekNb}{ The plate number}
    \item{RowNb}{ The row number}
    \item{ColNb}{ The column number}
    \item{ScreenNb}{ The screen number}
    \item{NbCells}{ E.g. the number of cells identified in the well (channel 2)}
    \item{PercCells}{ The ratio (number of identified cells)/(number of identified objects)}
    }
  }
  \item{expname}{A character string, assigning a name to the experiment. This will be used as title in the html output generated by \code{rnaither}.}
  \item{excludeCellcounts}{a string constant, one of \code{"none"}, \code{"lowest"}, \code{"both"}, \code{"lowestperplate"} or \code{"bothperplate"}. The default is \code{"none"}. This
    parameter can be used to exclude wells from the analysis that have very low or very high numbers of cells.
    \itemize{
    \item{"none"}{No wells will be excluded based on the number of cells they contain.}
    \item{"lowest", "lowestperplate"}{The wells with the lowest 5 percent of cellcounts will be excluded from further analysis. "lowest" will consider the entire screen at once, and
      exclude the wells that are overall the lowest 5 percent. "lowestperplate" will consider each plate separately, excluding on each plate the 5 percent of wells having the lowest cellcounts.}
    \item{"both", "bothperplate"}{The wells with the lowest and highest 5 percent of cellcounts will be excluded from further analysis. Excluding wells with high cell counts may be useful for image based screens, if it is suspected that cells overlap in images, which might cause problems for image processing. "both" will consider the entire screen at once, and
      exclude the wells that are overall the lowest and highest 5 percent. "bothperplate" will consider each plate separately, excluding on each plate the 5 percent of wells having the lowest and highest cellcounts.}
    }
  }
  \item{logtransform}{A logical variable, specifying whether or not the signal intensities should be log-transformed. Default is FALSE.}
  \item{normalization}{A list of strings containing the normalization steps to carry out. The default are is \code{c("lowess","bscore")}. The following normalization procedures are available:
  \itemize{
    \item{"lowess"}{To carry out lowess normalization. This corrects for effects of cell counts on the signal intensities.}
    \item{"liwong"}{To carry out Li-Wong rank normalization of the signal intensities.}
    \item{"varadjust"}{To divide each signal intensity value by the variance of the signal intensities on the respective plate.}
    \item{"divnorm"}{To divide each signal intensity value by the median signal intensity of the respective plate.}
    \item{"quantile"}{To carry out a quantile normalization on the signal intensities.}
    \item{"bscore"}{To carry out a bscore normalization on the signal intensities (corrects for spatial effects on a plate.}
    \item{"zscore"}{To carry out a zscore normalization (subtract median of plate, divide by median absolute deviation per plate).}
    \item{"negcontrol"}{To normalize on the negative controls - subtract median of negative controls, divide by MAD of negative controls, per plate.}
    \item{"percontrol"}{To do a percentage of controls normalization - Rescale signal intensities so that mean of negative controls is 100, mean of positive controls is 0.}
    \item{"percneg"}{To do a percentage of negative controls normalization - set mean of negative controls to 100, zero signal intensity remains at 0.}
    Normalization routines will be executed in the order as they occur in the list.
    }
  }
  \item{test}{Specify what statistical test should be used to identify hits. One of
  \itemize{
    \item{"ttest"}{to carry out a t-test if the mean score for a given siRNA / Gene is 0.}
    \item{"wilcox"}{to carry out a Wilcoxon test if the mean score for a given siRNA / Gene is 0.}
    \item{"none"}{to carry out no statistical test.}}
    The default is "ttest".
  }
  \item{scorethresh}{The threshold on the normalized score to be used to identify hits. The default is 2.0, hence siRNAs with score > +2 or score < -2 are considered hits.}
  \item{pvalthresh}{The threshold on the p-value from the statistical test to be used to identify hits. The default is 0.05}
  \item{dogo}{A logical variable, specifies whether or not a Gene Ontology-based analysis should be carried out. This parameter is currently ignored, GO is presently not supported nby
    the \code{rnaither} wrapper.}
  \item{outdir}{a string specifying the directory in which the results should be stored. Can be an absolute or relative path.}
  \item{layoutnames}{A list of strings, that can be used to assign names to different layouts in the screen. The list should contain the same number of elements as there are different layouts in the screen. These names will be used as labels for the layouts in the html output. If this parameter is not specified, layouts will be numbered in the canonical way.}
  \item{makeplots}{TRUE or FALSE, if set to FALSE, only a subset of the quality control plots will be generated. This speeds up processing, but will result in missing images in the html output.}
  \item{reorder}{logical variable, indicating whether dataset should be reordered prior to processing further. This is recommended if the data frame is incomplete, i.e. if wells or plates are missing completely. \code{reorder=T} will considerably slow down the analysis.}
}

\value{
Generates the html output file \code{index.html} in the directory specified by the \code{outdir} parameter.
}


\examples{
data(exampleDataset, package="RNAither")

rnaither(dataset, expname="Example", excludeCellcounts="none", logtransform=FALSE, normalization=c("lowess","zscore"), test="none", scorethresh=2.0, pvalthresh=0.05, outdir="results")

}

\keyword{ dplot }
\keyword{ device }
\keyword{ arith }
\keyword{ manip }
\keyword{ database }
\keyword{ methods }
\keyword{ univar }
\keyword{ datagen }
\keyword{ misc }
\keyword{ utilities }
