\name{scoredDataset1}
\docType{data}
\alias{scoredDataset1}
\title{A dataset containing an additional column showing the p-values, after a median normalization and a t-test}
\description{See \code{\link{divNorm}} and \code{\link{Ttest}} for details}
\usage{scoredDataset1}
\format{see \code{\link{generateDatasetFile}} for details}
\keyword{datasets}