\name{scoredDataset2}
\docType{data}
\alias{scoredDataset2}
\title{A dataset containing an additional column showing the p-values after a Mann-Whitney test}
\description{See \code{\link{MannWhitney}} for details}
\usage{scoredDataset1}
\format{see \code{\link{generateDatasetFile}} for details}
\keyword{datasets}